pyproject.toml
[project]
name = "audiotranscriber"
version = "0.0.3"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "deepgram-sdk>=4.8.1",
    "dotenv>=0.9.9",
    "ffmpeg>=1.4",
    "huggingface-hub>=0.32.3",
    "ipykernel>=6.29.5",
    "langchain-together>=0.3.0",
    "langdetect>=1.0.9",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "openai>=1.75.0",
    "openai-whisper>=20240930",
    "pyannote-audio>=3.3.2",
    "pyaudio>=0.2.14",
    "pycld2>=0.42",
    "scipy>=1.15.2",
    "soundcard>=0.4.4",
    "sounddevice>=0.5.1",
    "speechrecognition>=3.14.2",
    "streamlit>=1.44.1",
    "together>=1.5.8",
]

[tool.uv]
link-mode = "copy"
